requirements.txt
# Telegram Bot Framework
aiogram==3.4.1

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
aiosqlite==0.19.0

# Data Processing & Reports
pandas==2.1.4
openpyxl==3.1.2
xlsxwriter==3.1.9
reportlab==4.0.7

# Validation & Serialization
pydantic==2.5.2
pydantic-settings==2.1.0

# Logging
loguru==0.7.2

# Utilities
python-dotenv==1.0.0
pytz==2023.3
python-dateutil==2.8.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.1
isort==5.13.2
mypy==1.8.0 